Update typings, adapt structure
