+++ package.json
@@ -15,11 +15,10 @@
     "lint": "npm run lint.client && npm run lint.server",
     "lint.client": "gulp lint.client",
     "lint.server": "gulp lint.server",
-    "postinstall": "npm prune && npm run typings && npm run dist",
+    "postinstall": "npm prune && npm run typings.install && npm run dist",
     "start": "npm run https",
-    "typings": "npm run typings.client && npm run typings.server",
-    "typings.client": "rimraf ./src/client/typings/ && (cd ./src/client/ && typings install)",
-    "typings.server": "rimraf ./src/server/typings/ && (cd ./src/server/ && typings install)"
+    "typings": "typings",
+    "typings.install": "rimraf ./typings/ && typings install"
   },
   "repository": {
     "type": "git",
@@ -67,7 +66,6 @@
     "gulp-inline-ng2-template": "^2.0.3",
     "gulp-load-plugins": "^1.2.4",
     "gulp-nodemon": "^2.0.7",
-    "gulp-rename": "^1.2.2",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-tslint": "^5.0.0",
     "gulp-tslint-stylish": "^1.1.1",
@@ -76,6 +74,6 @@
     "html-minifier": "^2.1.3",
     "rimraf": "^2.5.2",
     "tslint": "^3.10.2",
-    "typings": "^0.8.1"
+    "typings": "^1.0.4"
   }
 }
