--- src/client/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "system",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "removeComments": true,
-    "noImplicitAny": true,
-    "outDir": "../../dist/client"
-  },
-  "exclude": [
-    "typings/main",
-    "typings/main.d.ts"
-  ]
-}
