--- src/client/typings.json
@@ -1,6 +1,0 @@
-{
-  "name": "twitchr-client",
-  "dependencies": {},
-  "devDependencies": {},
-  "ambientDependencies": {}
-}
