--- src/server/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "sourceMap": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "removeComments": true,
-    "noImplicitAny": true,
-    "outDir": "../../dist/server"
-  },
-  "exclude": [
-    "typings/browser",
-    "typings/browser.d.ts"
-  ]
-}
