--- src/server/typings.json
@@ -1,24 +1,0 @@
-{
-  "name": "twitchr-server",
-  "dependencies": {
-    "debug": "registry:npm/debug#2.0.0+20160211003958",
-    "glob": "registry:npm/glob#6.0.0+20160211003958",
-    "mime": "registry:npm/mime#1.3.0+20160423043021",
-    "mongodb": "registry:npm/mongodb#2.1.0+20160422172506",
-    "request": "registry:npm/request#2.69.0+20160428223725"
-  },
-  "devDependencies": {},
-  "ambientDependencies": {
-    "connect-mongo": "registry:dt/connect-mongo#0.0.0+20160317120654",
-    "express": "registry:dt/express#4.0.0+20160317120654",
-    "express-serve-static-core": "registry:dt/express-serve-static-core#0.0.0+20160322035842",
-    "express-session": "registry:dt/express-session#0.0.0+20160331200931",
-    "helmet": "registry:dt/helmet#0.0.0+20160501161936",
-    "irc": "registry:dt/irc#0.3.12+20160317120654",
-    "mongoose": "registry:dt/mongoose#3.8.5+20160505185535",
-    "morgan": "registry:dt/morgan#1.2.2+20160317120654",
-    "node": "registry:dt/node#4.0.0+20160505172921",
-    "serve-favicon": "registry:dt/serve-favicon#0.0.0+20160316155526",
-    "serve-static": "registry:dt/serve-static#0.0.0+20160501131543"
-  }
-}
