+++ src/typings/app-root-path.d.ts
@@ -1,55 +1,33 @@
 /*!
     twitchr - A twitch bot providing IRC based assistance
     Copyright (C) 2016  Jonas Bürkel
-
+    
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
     the Free Software Foundation, either version 3 of the License, or
     (at your option) any later version.
-
+    
     This program is distributed in the hope that it will be useful,
     but WITHOUT ANY WARRANTY; without even the implied warranty of
     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
     GNU General Public License for more details.
-
+    
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import {Client, IClientOpts} from 'irc';
-import {DefaultOptions} from 'connect-mongo';
+declare module 'app-root-path' {
+    module appRootPath {
+        export let path: string;
 
-declare module 'connect-mongo' {
-    interface DefaultOptions {
-        /**
-         * @default 0
-         */
-        touchAfter?: number;
+        export function require(pathToModule: string): Object;
+
+        export function resolve(pathToModule: string): string;
+
+        export function setPath(explicitlySetPath: string): void;
+
+        export function toString(): string;
     }
-}
 
-declare module 'irc' {
-    interface IClientOpts {
-        /**
-         * @default ''
-         */
-        password?: string;
-    }
-}
-
-declare global {
-    module Express {
-        export interface Session {
-            ircClient?: Client;
-            oauth?: Oauth2Token;
-            oauthState?: string;
-            originalUrl?: string;
-        }
-
-        interface Oauth2Token {
-            access_token: string;
-            refresh_token: string;
-            scope: string[];
-        }
-    }
+    export = appRootPath;
 }
