+++ typings.json
@@ -1,0 +1,23 @@
+{
+  "name": "twitchr",
+  "dependencies": {
+    "debug": "registry:npm/debug#2.0.0+20160511151334",
+    "glob": "registry:npm/glob#6.0.0+20160211003958",
+    "mime": "registry:npm/mime#1.3.0+20160423043021",
+    "mongodb": "registry:npm/mongodb#2.1.0+20160527161729",
+    "mongoose": "registry:npm/mongoose#4.3.5+20160531094808",
+    "request": "registry:npm/request#2.69.0+20160428223725"
+  },
+  "globalDependencies": {
+    "connect-mongo": "registry:dt/connect-mongo#0.0.0+20160317120654",
+    "express": "registry:dt/express#4.0.0+20160317120654",
+    "express-serve-static-core": "registry:dt/express-serve-static-core#0.0.0+20160602151406",
+    "express-session": "registry:dt/express-session#0.0.0+20160331200931",
+    "helmet": "registry:dt/helmet#0.0.0+20160512043842",
+    "irc": "registry:dt/irc#0.3.12+20160521154229",
+    "morgan": "registry:dt/morgan#1.7.0+20160524142355",
+    "node": "registry:dt/node#6.0.0+20160602155235",
+    "serve-favicon": "registry:dt/serve-favicon#0.0.0+20160316155526",
+    "serve-static": "registry:dt/serve-static#0.0.0+20160501131543"
+  }
+}
